refactor: move common to `common.py` file
