+++ dataclass_rest/errors.py
@@ -1,0 +1,6 @@
+class ApiError(Exception):
+    pass
+
+
+class NotFoundError(ApiError):
+    pass
